+++ Bot/Registration/ChatRegistration.cs
@@ -1,0 +1,16 @@
+using Lunaris2.Handler.ChatCommand;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace Lunaris2.Registration;
+
+public static class ChatRegistration
+{
+    public static IServiceCollection AddChat(this IServiceCollection services, IConfiguration configuration)
+    {
+        services.AddSingleton<ChatSettings>();
+        services.Configure<ChatSettings>(configuration.GetSection("LLM"));
+        
+        return services;
+    }
+}
